Create 9.xyzcentre in string
Given a string, does "xyz" appear in the middle of the string? To define middle, we'll say that the number of chars to the left and right of the "xyz" must differ by at most one. This problem is harder than it looks.
xyzMiddle("AAxyzBB") → true
xyzMiddle("xyAxyzBByz") → true
xyzMiddle("AxyzBBB") → false