Create Triplet that Sum to Given Value in Java
A triplet that sums to a given value is a set of three elements in an array whose sum is equal to the given value. For example, if we have an array [1, 2, 3, 4, 5] and the target sum is 9, then a triplet that sums to 9 could be (2, 3, 4) or (1, 2, 6), as both sets of elements add up to 9. The key requirement is that the sum of the three elements in the triplet is equal to the given value.